taicol v1.0

+++ app/Http/Controllers/MyNamespaceUsageController.php
@@ -4,18 +4,35 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\MyNamespaceCollection;
 use App\Http\Resources\PersonCollection;
+use App\Http\Resources\TaxonNameCollection;
+use App\Http\Resources\UsageCollection;
 use App\MyNamespace;
 use App\MyNamespaceUsage;
 use App\Person;
 use App\Reference;
+use App\TaxonName;
 use Illuminate\Http\Request;
-use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Validation\Rule;
 
 class MyNamespaceUsageController extends Controller
 {
+
+    public function index(Request $request, $namespaceId)
+    {
+        $namespace = MyNamespace::find($namespaceId);
+
+        if ($namespace->user_id !== $request->user()->id) {
+            return response()->json([], 401);
+        }
+
+        $usages = MyNamespaceUsage::where('namespace_id', $namespaceId)->orderBy('group')->orderBy('order')->get();
+
+        $data = MyNamespaceCollection::collection([$namespace])->first()->toArray($request);
+        $data['usages'] = UsageCollection::collection($usages);
+
+        return response($data);
+    }
 
     public function show($namespaceId, $usageId)
     {
@@ -35,38 +52,93 @@ class MyNamespaceUsageController extends Controller
             ->where('id', $usageId)
             ->first();
 
+        $acceptedUsage = MyNamespaceUsage::with([
+            'parent',
+            'taxonName.nomenclature',
+            'taxonName.rank',
+            'taxonName.authors',
+            'taxonName.exAuthors',
+            'taxonName.reference.authors',
+            'taxonName.originalTaxonName',
+            'taxonName.originalTaxonName.authors',
+            'taxonName.originalTaxonName.exAuthors',
+            'namespace'
+        ])
+            ->where('namespace_id', $namespaceId)
+            ->where('is_indent', false)
+            ->where('group', $usage->group)
+            ->where('id', '!=', $usage->id)
+            ->first();
+
+        $typeName =  $typeName = ($usage->properties['type_name'] ?? '') ? TaxonNameCollection::collection([
+            TaxonName::with([
+                'authors',
+                'exAuthors',
+                'reference',
+                'nomenclature',
+                'originalTaxonName.authors',
+                'originalTaxonName.exauthors'
+            ])->find((int) $usage->properties['type_name'])
+        ])[0] : null;
+
         return response([
             'id' => $usage->id,
             'namespace' => $usage->namespace,
-            'taxon_name' => $usage->taxonName,
-            'parent_taxon_name' => $usage->parent,
+            'taxon_name' => TaxonNameCollection::collection([$usage->taxonName])[0],
+            'parent_taxon_name' => $usage->parent ? TaxonNameCollection::collection([$usage->parent])[0] : null,
             'status' => $usage->status,
             'properties' => $usage->properties,
+            'type_name' => $typeName,
+            'group' => $usage->group,
             'per_usages' => collect($usage->per_usages)->map(function ($r) {
                 $r['target'] = isset($r['reference_id']) ? Reference::with('authors')->find($r['reference_id']) : null;
                 return $r;
             }),
             'type_specimens' => collect($usage->type_specimens)->map(function ($t) {
-                    $t['collectors'] = PersonCollection::collection(Person::whereIn('id', $t['collectors'])->get());
+                    $t['collectors'] = PersonCollection::collection(Person::whereIn('id', $t['collector_ids'])->get());
                     return $t;
                 }) ?? [],
             'name_remark' => $usage->name_remark,
             'custom_name_remark' => $usage->custom_name_remark,
+            'accepted_usage' => $acceptedUsage,
         ]);
     }
 
     public function update(Request $request, $namespaceId, $usageId)
     {
+        $taxonNameId = $request->get('taxon_name_id');
+
         $request->validate([
+            'status' => 'required',
+            'properties.indications' => 'required_if:status,misapplied,undetermined',
+            'parent_taxon_name_id' => function($attribute, $parentTaxonNameId, $fail) use ($taxonNameId) {
+                // validate parent taxon name id
+                $taxonName = TaxonName::with(['rank'])->where('id', $taxonNameId)->first();
+
+                $parentTaxonName = TaxonName::with(['rank'])->where('id', $parentTaxonNameId)->first();
+
+                $speciesParentKey = ['genus', 'subgenus', 'section', 'subsection'];
+
+                // 種下階層
+                $underSpecies = ['aberration', 'morph', 'stirp', 'race', 'special-form', 'subform', 'form', 'nothovariety', 'subvariety', 'variety', 'nothosubspecies', 'subspecies'];
+
+                if (!$taxonName) {
+                    $fail('請選擇正確上階層');
+                } else if ($taxonName->rank->key === 'species' && !in_array($parentTaxonName->rank->key, $speciesParentKey)) {
+                    $fail('請選擇正確上階層');
+                } else if (in_array($taxonName->rank->key, $underSpecies) && $taxonName->properties['species_id'] != $parentTaxonNameId) {
+                    $fail('請選擇正確的種下上階層');
+                }
+            },
             'type_specimens.*.use' => 'required',
             'type_specimens.*.kind' => 'required|integer',
             'type_specimens.*.country' => 'required_if:type_specimens.*.kind,1',
-            'type_specimens.*.specimens.*.herbarium' => 'required|min:1',
+            'type_specimens.*.specimens.*.herbarium' => 'required_if:type_specimens.*.kind,1',
             'type_specimens.*.collection_year' => 'max:4',
-            'type_specimens.*.collection_month' => 'max:2',
             'type_specimens.*.collection_day' => 'max:2',
-            'type_specimens.*.collectors' => 'required_if:type_specimens.*.kind,1|array|exists:persons,id',
+            'type_specimens.*.collector_ids' => 'required_if:type_specimens.*.kind,1|array|exists:persons,id',
             'type_specimens.*.isotypes.*.herbarium' => 'required',
+            'type_specimens.*.lecto_designated_reference' => 'required_if:type_specimens.*.use,lectotype',
             'per_usages.*.reference_id' => 'required',
             'per_usages.*.show_page' => 'integer|nullable',
         ], [
@@ -100,6 +172,7 @@ class MyNamespaceUsageController extends Controller
             $originUsage->custom_name_remark = $usage['custom_name_remark'] ?? '';
             $originUsage->properties = $usage['properties'] ?? [];
             $originUsage->per_usages = $usage['per_usages'] ?? [];
+
             $originUsage->save();
 
             DB::commit();
@@ -108,7 +181,7 @@ class MyNamespaceUsageController extends Controller
             ]);
         } catch (\Exception $e) {
             DB::rollBack();
-            Log::error("[usage] - udpate::fail $e->getMessage()");
+            Log::error("[usage] - udpate::fail {$e->getMessage()}");
             return response()->json([
                 'message' => 'error'
             ], 500);
@@ -123,11 +196,11 @@ class MyNamespaceUsageController extends Controller
             '*.type_specimens.*.use' => 'required',
             '*.type_specimens.*.kind' => 'required|integer',
             '*.type_specimens.*.country' => 'required_if:type_specimens.*.kind,1',
-            '*.type_specimens.*.specimens.*.herbarium' => 'required|min:1',
+            '*.type_specimens.*.specimens.*.herbarium' => 'required_if:type_specimens.*.kind,1',
             '*.type_specimens.*.collection_year' => 'max:4',
-            '*.type_specimens.*.collection_month' => 'max:2',
             '*.type_specimens.*.collection_day' => 'max:2',
-            '*.type_specimens.*.collectors' => 'required_if:type_specimens.*.kind,1|array|exists:persons,id',
+            '*.type_specimens.*.collectors.*.id' => 'exists:persons,id',
+            '*.type_specimens.*.collectors' => 'required_if:type_specimens.*.kind,1|array',
             '*.type_specimens.*.isotypes.*.herbarium' => 'required',
         ], [
             'min' => '必填',
@@ -145,45 +218,52 @@ class MyNamespaceUsageController extends Controller
             return response()->setStatusCode(404);
         }
 
-        $namespace->touch();
+        try {
+            $namespace->touch();
 
-        $group = 0;
-        foreach ($usages as $index => $usage) {
-            $isTitle = (bool) ($usage['is_title'] ?? false);
+            $group = 0;
+            foreach ($usages as $index => $usage) {
+                $isTitle = (bool) ($usage['is_title'] ?? false);
 
-            if (!$usage['is_indent']) {
-                $group += 1;
-            }
-
-            if (isset($usage['id']) && $usage['id']) {
-                $currentUsage = MyNamespaceUsage::find($usage['id']);
-                if ($currentUsage && isset($usage['is_deleted'])) {
-                    $currentUsage->delete();
-                    continue;
+                if (!$usage['is_indent']) {
+                    $group += 1;
                 }
-            } else {
-                $currentUsage = new MyNamespaceUsage();
-                $currentUsage->namespace_id = $namespaceId;
+
+                if (isset($usage['id']) && $usage['id']) {
+                    $currentUsage = MyNamespaceUsage::find($usage['id']);
+                    if ($currentUsage && isset($usage['is_deleted']) && (bool) $usage['is_deleted']) {
+                        $currentUsage->delete();
+                        continue;
+                    }
+                } else {
+                    $currentUsage = new MyNamespaceUsage();
+                    $currentUsage->namespace_id = $namespaceId;
+
+                    $currentUsage->parent_taxon_name_id = $usage['parent_taxon_name_id'] ?? null;
+                    $currentUsage->is_for_publish = false;
+                    $currentUsage->status = $usage['status'] ?? '';
+                    $currentUsage->type_specimens = $usage['type_specimens'] ?? [];
+                    $currentUsage->name_remark = $usage['name_remark'] ?? '';
+                    $currentUsage->custom_name_remark = $usage['custom_name_remark'] ?? '';
+                    $currentUsage->properties = $usage['properties'] ?? [];
+                    $currentUsage->per_usages = $usage['per_usages'] ?? [];
+                    $currentUsage->taxon_name_id = (int) $usage['taxon_name_id'];
+                }
+
+                $currentUsage->group = $group;
+                $currentUsage->order = $index;
+
+                $currentUsage->is_title = $isTitle;
+                $currentUsage->is_indent = (bool) $usage['is_indent'];
+                $currentUsage->save();
             }
 
-            $currentUsage->parent_taxon_name_id = $usage['parent_taxon_name_id'] ?? null;
-            $currentUsage->is_for_publish = false;
-            $currentUsage->status = $usage['status'] ?? false;
-            $currentUsage->type_specimens = $usage['type_specimens'] ?? [];
-            $currentUsage->name_remark = $usage['name_remark'] ?? '';
-            $currentUsage->custom_name_remark = $usage['custom_name_remark'] ?? '';
-            $currentUsage->properties = $usage['properties'] ?? [];
-            $currentUsage->per_usages = $usage['per_usages'] ?? [];
-            $currentUsage->taxon_name_id = (int) $usage['taxon_name_id'];
-            $currentUsage->group = $group;
-            $currentUsage->order = $index;
-
-            $currentUsage->is_title = $isTitle;
-            $currentUsage->is_indent = (bool) $usage['is_indent'];
-            $currentUsage->save();
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollBack();
         }
 
-        DB::commit();
+
         return response([]);
     }
 }
